+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-- Adding support for Amazon regions
+- Adding support for Amazon locales support
 
 ## [0.2.0]
 ### Added